refactor(vm): move Restore/crash recovery/watchdog onto Manager
Phase 2e of vm-lifecycle-extraction. Restore + reconnect, the QEMU exit
crash handler chain, and the pending-launch watchdog leave the daemon
package and become methods on vm.Manager.

- vm/restore.go — Manager.Restore loads per-node state, classifies VMs
  (terminate / stop migrate to shared KV; running with live QEMU and
  valid NBD sockets reconnect via QMP; transitional states finalise),
  and fans Run out across the relaunch queue with the existing
  semaphore.
- vm/crash_recovery.go — Manager.HandleCrash, MaybeRestart, and
  RestartCrashedInstance own the post-exit chain. ClassifyCrashReason
  and RestartBackoff become exported pure helpers.
- vm/watchdog.go — Manager.StartPendingWatchdog spawns the
  stuck-in-pending scanner. Daemon.Start drives it with d.ctx.
- ManagerHooks gains OnInstanceRecovering, fired by Restore once per
  relaunch so the daemon can early-subscribe ec2.cmd.<id> while a
  concurrent terminate races; OnInstanceUp on launch success
  reinstalls both per-instance subs idempotently.
- Deps gains ConsumeCleanShutdownMarker so the manager doesn't reach
  into JetStream for the marker decision.
- ResourceController gains CanAllocate so MaybeRestart can precheck
  capacity without round-tripping the EC2 SDK type.

Daemon-side wrappers (handleInstanceCrash, maybeRestartInstance,
restoreInstances, areVolumeSocketsValid, classifyCrashReason,
restartBackoff) are preserved as one-liner shims so the existing
health/state test cases keep their bare-name calls; phase 2f migrates
those into vm/ and deletes the shims.

Daemon.Start: vmMgr.Restore() runs before subscribeAll, with
OnInstanceUp/OnInstanceRecovering hooks firing on the same NATS
connection that connectNATS already established. vmMgr.StartPendingWatchdog
replaces the pre-2e startPendingWatchdog goroutine.

Net change: -486 lines on the daemon side. make preflight clean.
